Test SnapshotStore saves the given EntitySnapshot
